coordinator: close evidence race, APNs token floor, and scheduler modularity findings
- Application-evidence grants validate the release-policy generation
  atomically under the registry lock: a grant derived from a stale
  snapshot is refused and triggers the same immediate re-challenge as a
  sweep invalidation, closing the clear/derive/grant race.
- Application evidence no longer requires an APNs device token;
  tokenless providers derive and keep evidence, and token possession is
  enforced solely by the APNs code-identity gate with its existing
  grace semantics.
- Evidence carry-forward re-proves the full release runtime fact set
  (python hash, runtime hash, metallib, pinned template hashes),
  failing closed when any field cannot be proven unchanged.
- APNs code-attest admission adds a durable per-SE-key push floor with
  bounded per-key budget rows (newest 8 + floor sentinel), so token
  churn on reconnect can no longer bypass the per-device floor or grow
  durable state unboundedly; genuine heartbeat rotations still lift the
  floor.
- mdm_scheduler.go split verbatim into focused config/queue/exec/
  callbacks/metrics units behind a thin orchestrator.
